\name{pplace}
\alias{pplace}
\docType{data}
\title{
A placement object as obtained with the read_sqlite function
}
\description{
A placement object as obtained with the read_sqlite function. In this example, a set of 100 sequence reads are placed over a 16S phylogeny. This example is a subset of those available for download at http://fhcrc.github.io/microbiome-demo/
}
\usage{data("pplace")}
\references{
http://fhcrc.github.io/microbiome-demo/
}
\examples{
data(pplace)
str(pplace)
}

