\name{print.pplace}
\alias{print.pplace}
\alias{print.jplace}
\title{
Compact display of pplace and jplace objects
}
\description{
Compact display of pplace and jplace objects
}
\usage{
\method{print}{pplace}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
a pplace or jplace object
}
  \item{...}{
further arguments passed to or from other methods
}
}
\author{
pierre lefeuvre
}
\examples{
data(pplace)
print(pplace)
}
