\name{BoSSA-package}
\alias{BoSSA-package}
\alias{BoSSA}
\docType{package}
\title{
\packageTitle{BoSSA}
}
\description{
\packageDescription{BoSSA}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{BoSSA}
\packageIndices{BoSSA}
BoSSA contains functions to read and plot phylogenetic placement files obtained using softwares such as pplacer, guppy, EPA and RAPPAS.
}
\author{
\packageAuthor{BoSSA}
Maintainer: \packageMaintainer{BoSSA}
}
\references{
- pplacer and guppy
http://matsen.fhcrc.org/pplacer/
http://matsen.github.io/pplacer/
- EPA
https://sco.h-its.org/exelixis/web/software/epa/index.html
- RAPPAS
https://gite.lirmm.fr/linard/RAPPAS
- Common file format
http://journals.plos.org/plosone/article?id=10.1371/journal.pone.0031009
}
