\name{pplace}
\alias{pplace}
\docType{data}
\title{
a placement object as obtained with the read_sqlite function
}
\description{
a placement object as obtained with the read_sqlite function. In this example, a set of 17 sequence reads are placed over a mastrevirus phylogeny. At the end of the read name, a number indicates the number of reads it represents (i.e. these reads are centroids of read clusters. the number indicate the size of the cluster).
}
\usage{data("pplace")}
\examples{
data(pplace)
str(pplace)
}

