\name{print.pplace}
\alias{print.pplace}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Compact display of pplace object
}
\description{
Compact display of pplace object
}
\usage{
\method{print}{pplace}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
a pplace object
}
  \item{...}{
further arguments passed to or from other methods
}
}
\author{
pierre lefeuvre
}
\examples{
data(pplace)
print(pplace)
}
